fix(macos): from_chars float/double availability + uint64_t/size_t type mismatch
- setup.py: add explicit -mmacosx-version-min=13.3 compile flag to ensure
  std::from_chars/to_chars floating-point overloads are available even when
  MACOSX_DEPLOYMENT_TARGET env var doesn't propagate through cibuildwheel
- reader.hpp, file_codec_stream001.h, file_codec_packet001.h,
  file_codec_stream_lz4_001.h: change rowLen from size_t to uint64_t to
  match vleDecode<uint64_t> template parameter (on ARM64 macOS, size_t is
  unsigned long but uint64_t is unsigned long long — same size, different
  types, so reference binding fails)

+++ include/bcsv/codec_file/file_codec_packet001.h
@@ -136,7 +136,7 @@ public:
         }
 
         // Read row length (VLE with checksum update)
-        size_t rowLen = 0;
+        uint64_t rowLen = 0;
         try {
             vleDecode<uint64_t, true>(is, rowLen, &packet_hash_);
         } catch (...) {
@@ -169,7 +169,7 @@ public:
                 + std::to_string(rowLen) + " > " + std::to_string(MAX_ROW_LENGTH) + ")");
         }
 
-        read_buffer_.resize(rowLen);
+        read_buffer_.resize(static_cast<size_t>(rowLen));
         is.read(reinterpret_cast<char*>(read_buffer_.data()),
                 static_cast<std::streamsize>(rowLen));
         if (!is || is.gcount() != static_cast<std::streamsize>(rowLen)) {
@@ -177,7 +177,7 @@ public:
         }
         packet_hash_.update(read_buffer_);
 
-        return std::span<const std::byte>(read_buffer_.data(), rowLen);
+        return std::span<const std::byte>(read_buffer_.data(), static_cast<size_t>(rowLen));
     }
 
     // ── Boundary / state signals ────────────────────────────────────────
